resolve conflict done
