+++ src/main/java/com/pbl6/VehicleBookingRental/user/domain/bus_service/Bus.java
@@ -1,35 +1,43 @@
 package com.pbl6.VehicleBookingRental.user.domain.bus_service;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.JoinTable;
+import jakarta.persistence.ManyToMany;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.PrePersist;
+import jakarta.persistence.Table;
+import lombok.Getter;
+import lombok.Setter;
 import java.util.List;
+import java.time.Instant;
 
 @Entity
-@Table(name = "bus")
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
+@Table(name="buses")
+@Getter
+@Setter
 public class Bus {
-
     @Id
-    private int id;
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
     private String licensePlate;
+    private Instant createAt;
 
+    @ManyToMany(fetch = FetchType.LAZY)
+    @JoinTable(name="bus_utilities", joinColumns = @JoinColumn(name = "bus_id"), inverseJoinColumns = @JoinColumn(name ="utility_id"))
+    private List<Utility> utilities;
 
     @ManyToOne
     @JoinColumn(name = "bus_type_id")
     private BusType busType;
 
+    @PrePersist
+    public void handleBeforeCreated(){
 
-    @ManyToOne
-    @JoinColumn(name = "bus_partner_id")
-    private BusPartner busPartner;
+        this.createAt = Instant.now();
+    }
 
-    @OneToMany(mappedBy = "bus",cascade = CascadeType.ALL)
-    @JsonIgnore
-    private List<BusUtilities> busUtilities;
 }
