+++ src/main/java/com/pbl6/VehicleBookingRental/user/domain/bus_service/BusType.java
@@ -1,22 +1,28 @@
 package com.pbl6.VehicleBookingRental.user.domain.bus_service;
 
 import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+import java.util.List;
 
 @Entity
-@Table(name = "bus_type")
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
+@Table(name="bus_types")
+@Getter
+@Setter
 public class BusType {
-
     @Id
-    private int id;
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
     private String name;
     private int numberOfSeat;
     private String chairType;
+
+    @OneToMany(mappedBy = "busType", fetch = FetchType.LAZY)
+    private List<Bus> buses;
 }
