+++ src/main/java/com/pbl6/VehicleBookingRental/user/util/FormatRestResponse.java
@@ -42,10 +42,4 @@ public class FormatRestResponse implements ResponseBodyAdvice<Object>{
         }
         return res;
     }
-
-    
-
-    
-
-    
 }
