+++ src/main/resources/application.properties
@@ -22,14 +22,6 @@ spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 #config jwt
 pbl6.jwt.base64-secret=1brevEpqtKzJgL1e3qAO3cnDnzVyvd9LoMiLeCe3dm2e9oGXD5MciszRjhNTCasKScibgrz4SCiqBNuSBRcaTA==
 #expiration: 10 days
-
-pbl6.jwt.token-validity-in-seconds=864000
+pbl6.jwt.access-token-validity-in-seconds=864000
+pbl6.jwt.refresh-token-validity-in-seconds=864000
 ###
-#spring.flyway.enabled=false
-#spring.flyway.baseline-on-migrate=true
-#spring.flyway.url=jdbc:mysql://localhost:3306/vehicle_booking_rental
-#spring.flyway.user=vehicle_booking_rental
-#spring.flyway.password=vehicle_booking_rental
-#spring.flyway.baseline-description="init"
-#spring.flyway.baseline-version=0
-
